Do not store cross compiler in config
The cross compiler variable should not be set in the config as can be
seen in commit 5d64b5826890 ("Add detection for a cross compiler").
It should never been there and went unnoticed. Remove it and force it
to be set via the environment, which we do and is expected.

EM-749

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>