Make the buildscript give a notice if makeflags are not set
When building on the local machine, one may want to use more then one
compilation thread to speed up compilation. This can be set via the
MAKEFLAGS environment variable, so let the user know that it was unset
when calling the build script.

Note: Do not set MAKEFLAGS in the default shell, as they apply to all
invocations of make. Also do not abuse this to have the build-slave
build with more jobs, as it is distributive to others.

EM-749

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>