Switch to 4.14 LTS
This is part 1 of our switch to the 4.14 LTS kernel. We load more
drivers as modules, that do not strictly have to be in the kernel to
reduce our kernel size.

Additionally, it should be noted that newer kernels require a different
devicetree pincontroller configuration and have some regulator
references renamed, which is also part of this patch as we can otherwise
not compile the kernel + devicetree.

EM-1920

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>