Ensure there is 3.3 volt on port G
Port G is powered by LDO4, but LDO4 is set to 2.8 volt by default.

Ensure there is at least enough power on LDO4 available.

Note that currently this is only a cosmetic/futureproofing fix.
U-Boot is currently responsible for setting the initial voltage of
LDO4. To fix this, a consumer is needed on the IO controller.

EM-1631

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>