bootscript: prevent binding of fbcon
Normally, the framebuffer console (fbcon) binds to the first available
framebuffer (for example fbcon=map:0). However this can cause issues
for slow displays such as the ssd1306. Here the at some point the
ssd1306 is still busy sending the buffer to the screen, more data is
generated to be sent to the screen (via the console). This causes an
error, that the device is busy sending data to the screen, which also
gets added to the queue of items to be sent to the screen. This then
continues forever.

By ensuring the framebuffer console is only mapped when needed
(currently only the installer maps it) we can better disable the
framebuffer console. By mapping it to an unused framebuffer, the
console gets disabled, so lets take the last possible framebuffer.

Note, this workaround will fail if we have 64 displays.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>