dts: Reduce programming station clock freq
The hdmi-i2c clock frequency of the display programming stations
sometimes fails due to wrongfully specced parts in combination with an
empty EDID.

The empty EDID forces some default timing on the HDMI peripheral which
in turn results in a i2c clock frequency that causes problems
electrically due to component choices.

For now, the clock is lowered to ensure we can program our eeprom's.
This will need to be revisited when the touchscreen test application
will also use this bus.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>