initramfs: Use watchdog to reboot
Busybox's reboot command doesn't actually trigger a reboot. Our entire
SoC actually doesn't support reboot/reset. The watchdog is used to
actually reboot the actual system.

So let us do this as well, use a super fast watchdog to reset the SoC
when a reboot is requested. If this still happens to fail (no
/dev/watchdog available for example) try a shutdown instead.

Contributes to issue: EM-2098
Closes issue: EM-2398

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>