Add support for rfkill on our usb hosts
This patch enables rfkill to be used on our usb hosts. Currently,
rfkill is not the designed framework for this, as it is explicitly
described for radio-frequency devices, but until we have a devkill or
similar device for this purpose, it will have to 'abuse' rfkill. This
abuse happens only in the devicetree. The kernel itself just see's a
wifi gpio.

Contributes to issue: EM-2406

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>