initramfs: Add support for rootdelay
The kernel supports the rootdelay paramater to yield time for devices
to settle before trying to mount them. While this parameter is ignored
by the kernel if we use an initramfs, we honor it the same.

Note however, that we sleep during the start of init, not specifically
before mount, as we also need it to search for updates.

Contributes to issue: EM-2098

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>