initramfs: Set default rootdelay to 2s
Sleep for 2 seconds during boot to allow the USB stack to settle
probing. This allows us to search for updates on USB storage, as
the init script is already started by the kernel before it has finished
probing all devices.

Note that this number is somewhat arbitrary and querying uevents may be
more reliable.

Contributes to issue: EM-2098

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>