ci: Use Dockerfile for automated builds
To make building of this package easier, docker is used.
This thus now requires a docker enabled gitlab-runner.

The CI job is split up into two basic parts. The first half
creates the docker image, if the keyword 'dockerfile: ' was
at the start of the subject. Only if that was successfully
pushed to a docker registry, the actual build continues.

Contributes to issue EMP-141

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>