Fixed docker build output permissions
The build output dir is created in the current working directory by the user executing docker_build,
then the build output dir is passed to docker as a volume and the dockers entry point is run
with the calling users UID. This way it is possible to have the build generated fails available when
the docker run is complete for archiving the deb package, the build output can be deleted because they
are created withe the correct UID and incremental builds for local build are made possible. A requirement
for this was that all the generated files are generated in the build outpt directory.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>