initrd: No longer focus on a toolbox
Initially, the plan was to have a toolbox, and perform the update from
within the toolbox. This required two files to be present, the toolbox
and the update.

This has now changed to only supply one file, the update, as a squashfs
image. Within the image, we have the toolbox still.

As such, we thus now only mount an update file and execute a script
within. The script within can do whatever it wants, including for
example mount the toolbox and perform the update using the tools from
the update.

Contributes to EMP-488

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>