init: Ensure init is actually an executable
As we sometimes may pass an invalid init or have a corrupt filesystem,
it is ever so useful to check this binary before attempting to boot it.
If we don't, the kernel will panic about a missing init. Now we will
drop to our rescue shell and can figure out whats going on or fix it.

Closes EMP-291.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>