dts: Remove spidev, we bitbang anyway
Initially we wanted to use the userspace spidev to program the
mainboard. While this is possible, our software cannot do this yet.
Especially as we export the pins as gpio's, the spi mux gets
overwritten.

We should do neither however and just use remote-proc for this, thus
lets remove spidev altogether.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>