Fix error in kernel build because of broken download URL.
The kernel build was failing because the busybox version URL is broken.
It has been updated to a new version. For now implemented a workaround
with a new version URL and at least notify the user if something is
wrong with the download.
An issue is logged to implement a proper fix EMP-20.

This issue addresses EMP-538.

version is now: busybox-static-1.28.4-r3

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>