Added usb2512b USB hub configuration I2C interface
In order to get the on-carrier usb2512b USB hub to work, it needs to be configured by the driver (linux/drivers/usb/misc/usb251xb.c, CONFIG_USB_HUB_USB251XB) during boot. To achive this, the hub is added to the msc-sm2-imx6x.dtsi device-tree.
This driver requires to reset the usb2512b USB hub prior to sending its configuration. To achive this, pad MX6QDL_PAD_GPIO_0, GPIO1_IO00 is repurposed as reset-gpios pin.

Please note: The latter is a violation of the SMARC2 standard and hence needs to be reverted once a SMARC2 GPIO pin is assigned as reset-gpios pin in the next board revision.

Contributes to: EMP-559