Fixed initramfs kernel module dependency issue
On a freshly checked out 'um-kernel' instance (i.e. what happens when
releasing a package), the proper dependent Kernel modules where not
added to initramfs because the modules.dep file was not generated in
that stage.

Now the modules are correctly build and installed in the
'kernel_build_modules' step for other steps to use properly.

In order to prevent the 'kernel_build_modules' build step to be executed
multiple times, rework of the 'deb_build' step was required. Right now
each step will output it's artifact to the 'DEBIAN_DIR' in the
'BOOT_FILE_OUTPUT_DIR' so that the 'deb_build' step can do some
sanity checking on the existence of the required files and package
them as the result.

Improved touched code by making variable naming more consistent and
preventing globbing and word splitting by adding double quotes to
variables.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>

Addresses: EMP-607, contributes to EMP-479