Fix build step dependecy issues on clean build
The issue addressed here is a 'chicken vs egg' problem, i.e. the Kernel modules depend on
the Kernel being build, the Kernel depends on the initramfs being prepared and the initramfs
depends on the modules being properly build and installed.

This was not a problem before because we did not pull in any modules into the initramfs that had any dependency,
and the Kernel modules dependency file is only completely generated when the complete Kernel including the modules
are build.

Now the Kernel build step will first prepare the initial initramfs, then build the Linux src including the modules,
then install the modules and create the dependency file, then add modules and dependant modules to the initramfs
and finally build the uImage Linux Kernel including the initramfs.

Addresses EMP-607

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>