EMP-658: Reassigned RESET_OUT# back to generic carrier board reset
Previously GPIO bank 7 pin 13 was assigned as a dedicated reset signal
to reset the I2C GPIO expander. This also required a modification in the
Kernel to actually operate I2C GPIO expander signals, as it does not do
this per default.

However, this needed to be reverted because GPIO bank 7 pin 13 carries
the RESET_OUT# signal. It is not a dedicated I2C GPIO expander reset
signal. It _does_ reset the I2C GPIO expander, but only as a side-effect
of resetting the whole carrier board.

Solves: EMP-658