Add trigger stage to CI
... and send variables to triggered job

... and add curl

Triggering is done by an `after_script` because the triggered script
wants to know which job ID triggered it.

Having a dedicated stage is more desirable, and this can be reverted
once we have a final decision on which system to use for storing
packages.

Contributes to EMP-674

Signed-off-by: martijnbruinenberg <m.bruinenberg@ultimaker.com>

Mask build trigger token