Minor improvements to shell scripts
Security: always exit the script
Logic: Simplify `if` statement
Logic: Don't push tagged image to registry if there are no changes

Other:

When building locally, run docker with -it so you can interact with or
cancel the build if you'd like.

Use shellcheck-alpine to shave off a few seconds.

Contributes to EMP-645 and EMP-675

Signed-off-by: martijnbruinenberg <m.bruinenberg@ultimaker.com>