Fix error with target update device when running from restore img
Normally in the um-kernel initrd process, the device where the um-update.swu image
is found is passed as the target update storage device. Altough this is correct for
a firmware update, it does not work for a firmware restore. When restoring we want to
update the eMMC instead of the MMC.

Unfortunately it is not trivial to determine what storage device is the SD card and
which is the internal flash. Therefore the 'recovery_data' partition is introduced
to serve as a recognition point to know if we are restoring or updating.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>