Always update mmcblk2 for msc-sm2-imx6dl based configurations.
On the msc som the mmc devices are always indexed in the same
way, it does not depend on where it boots from.

MMC0 : SD slot on msc-sm2-imx6dl SOM
MMC1 : Ultimainboard SD slot
MMC2 : eMMC internal flash.

Adapted the interface to jedi-system-update to pass in the display type
instead of the article number.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>