EMP-781: Configured-in Micrel specific Phy driver
Before, the Kernel was using the Generic Phy driver:
[    7.154739] Generic PHY 2188000.ethernet-1:00: attached PHY driver [Generic PHY] (mii_bus:phy_addr=2188000.ethernet-1:00, irq=POLL)
The MSC reference board is using a Micrel specific driver through the
Freescale FEC Phy driver:
[    7.414773] fec 2188000.ethernet eth0: Freescale FEC PHY driver [Micrel KSZ9031 Gigabit PHY] (mii_bus:phy_addr=2188000.ethernet:00, irq=-1)
Since the MSC SoM has a Micrel KSZ9031 Phy on board, it is recommended
to use the corresponding specific driver over the generic driver.
This commit configures-in the Micrel specific Phy driver, capable of
driving a wide variety of Micrel Phys, including the KSZ9031.
With this configured in, Kernel logging looks different from the MSC
board, but that's due a newer Kernel on our side:
[    7.193183] Micrel KSZ9031 Gigabit PHY 2188000.ethernet-1:00: attached PHY driver [Micrel KSZ9031 Gigabit PHY] (mii_bus:phy_addr=2188000.ethernet-1:00, irq=POLL)
Functionality has been verified to be idential.

IMPORTANT: With this configuration, the board has been seen rebooting
upon loading this device driver. This is most likely caused by a mis-
configured LDO, powering it. This problem still needs to be addressed.
If this occurs, the board will be temporarilly stuck in a boot-loop,
but will eventually boot. When it makes it through the bool-loop, no
unexpected resets have been observed.