Be backwards compatible with 5.2.x jedi-system-update
The dependency of the article number being passed by U-Boot
to determine the display type has been removed and the
restore mode of the update tooling was not yet implemented.
therefore requiring an interface change in jedi-system-update.

This commit makes sure it does not brake when downgrading
back to 5.2.x versions.

Contributes to EMP-850

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>