Fix symbols not found while loading framebuffer modules in initrd
This is not a problem, it is just the modprobe tool telling us
that its missing dependencies. These are in turn resolved impplicitely
by depmod, making sure the dependent modules are loaded. This is
only visible because we use modprobe with the verbose flag.

This is fixed by making the module inclusion explicit and in the
right order.

Fixes EMP-877.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>