Add a smarter more generic keybord detection with Python script.
We can now listen for all input devices ad filter keyboards.
The script will check if developer mode is enabled, if not
Okuda will keep running.
If the gpm (General purpose mouse) package is installed, it will
setup the mouse to be used in the framebuffer console.

Contributed by: Jack Ha.

Signed-off-by: Raymond Siudak <r.siudak@ultimaker.com>