E1-989 Show a blinking LED at the end of the update/restore procedure
The Olimex board has a green LED. When the update/restore procedure has finished,
we'll make this LED blink in a pattern of fast-fast-pause to indicate to the
user that we are ready. This has the advantage that the feedback works, even
without attached display.

During the update, the LED will flash on access to the eMMC program memory.