SEM-146 Write country lock during EEPROM provisioning
Extend the printer provisioning logic in the `init.sh` boot script, so
that a "locked country code" can be set in the mainboard EEPROM.

The existing provisioning code looks for an `article_number` file on a
USB drive that's inserted during boot (ext4, single partition). The
contents of that file, a single line of text in hex notation describing
the BOM number, is then written to the mainboard EEPROM using
`i2ctransfer`.

We duplicate this code, so that an additional file `country_code_lock`
can be used to also write a "locked country code" to the EEPROM, at
address 0x0118. This file is optional, so default provisioning keeps
whatever data is here, intact. The file works the same way: it should
contain a single line of text, with a hex representation of the locked
country code.

Co-authored-by: Onno Broekmans <o.broekmans@ultimaker.com>
Signed-off-by: Robin den Hertog <r.denhertog@ultimaker.com>