SEM-146 Boyscout: improve init.sh script
Improve the code in the `check_and_set_eeprom_data()` part of the
`init.sh` boot script. Based on feedback from @alexborro:

- No longer let the while loop span the entire code within the function.
If you analyze the purpose and flow of the while loop, it turns out it's
only ever looping to wait for the USB drive to show up under the `/dev/`
directory. All other code paths lead to either a `return` or the end of
the while loop. That makes things unnecessarily confusing.

- If an `i2ctransfer` call fails, don't exit the function completely,
but give the *other* provisioning file another chance. Do make sure that
all failures and skips are properly logged.

Co-authored-by: Alex Borro <alexborro@gmail.com>