CES-2886: Enable Touchscreen for Supersonic
Since we are no longer using the EEPROM EDID data, the lvds configuration had to be updated. The EDID data we have in the edid repo was not working for this board. To fix the offset it had, I calculated and guesstimated the best values since we don't have the display datasheet available.

This board also uses a different touchscreen driver, and a different i2c gpio to drive the reset signal. Those were updated too, to follow the schematics in Altium.