[CES-2811] Enable rtw88 wifi support and copy binaries
With this commit support in the kernel for rtw88 devices, like the
RealTek 8822CE wifi chipset that we use as a PCI WiFi module is enabled.

Enabling kernel support must now also go hand in hand with the
proprietary firmware binaries for the RTL8822CE hardware. This is a
change since Kernel 4.14.
So these binaries are now copied from the linux firmware source and
stored under version control within our 'proprietary_firmware'
directory.

This is similar to what has always been donefor the IMX SDMA binaries.

In the future we will probably have to add more proprietary firmware to
our kernel package, so we also adjust to function which copies the
binaries to now just copy everything we place in the
'proprietary_firmware' directory to the /lib/modules/ dirextory of the
um-kernel debian package.

In order to copy everything except a README.md file, rsync is now added
to the docker file - this allows us to copy files but also exclude
specific files in a dir.