[CES-3178] The S6 camera appears to have power-up issues when the board is powered on with the camera connected. This was addressed in hardware (see ELS-458), but introduced a worse side effect: the S8 camera now causes the i.MX8 USB Host Controller to freeze during boot (see ELS-462).
A workaround involves configuring the Microchip USB Hub (2514B) reset pin as a GPIO hog with a pull-up.
Pay attention the signal is inverted when connected to the hub. During kernel boot, the GPIO pin is first pulled
high by the pull-up, then driven low by the hog configuration. This toggling of the hub’s reset line forces a
re-enumeration of the USB devices and resolves the issue.