[CES-3178] The S6 camera appears to have power-up issues when the board is powered on with the camera connected. This was addressed in hardware (see ELS-458), but it introduced a worse side effect: the S8 camera now causes the i.MX8 USB Host Controller to freeze during boot (see ELS-462).
The root cause appears to be the time the S6 camera takes to power up. The Microchip USB2514B hub, when operating
in standalone mode (not configured via I2C), uses a default POWER-ON TIME of 100 ms. This is the delay the hub
waits before attempting to enumerate devices after enabling power to the port via the PRTPWR pins. For the S6 camera,
100 ms does not seem sufficient for it to be fully ready for enumeration, leading to failures.

Connecting the USB2514B to the I2C bus and increasing the POWER-ON TIME to 110 ms resolves the issue. In this configuration,
we set it to 200 ms to be on the safe side.