[CES-3322] Set NGP Proto screens to the correct touch orientation
With this PR, if a printer has the BOM number 109*, then it is an NGP
printer and uses a large Colorado like screen, except it is not an
inverted screen like the production printers.

So we use a different dts for machines with BOM's 109* which is very
similar to the ulticontroller4.0-lvds-1024x600.dts, just the inversion
of the x and y is removed.